Force use of IPv4 instead of IPv6 (-Djava.net.preferIPv4Stack=true)
Notably, this prevents tests to fail on Mac OS X because of timeouts caused
by resolution of localhost and IPv6.  Storm uses a hardcoded test timeout of
5 seconds (see `TEST-TIMEOUT-MS` in `backtype.storm.testing.clj`), which is
often exceeded because of the long duration of resolution on Mac OS X
when IPv6 is enabled.

For more details on the timeout issue see:
https://issues.apache.org/jira/browse/STORM-70
https://issues.apache.org/jira/browse/ZOOKEEPER-1476