Squelch error messages logged by org.apache.curator.framework.imps.CuratorFrameworkImpl
These messages are logged (directly or indirectly) because of ZK timeouts
in storm-kafka, i.e. Storm's built-in KafkaSpout we are using to read
Kafka data into our Storm topologies.  It is unclear whether we can
prevent those messages in the first place as we are not certain what
causes them.

This commit is a follow-up on commit b4a9bc0, which disables ZK retries
for our Kafka/Storm test topology.